Reorganize global dialogs
