+++ src/app/dashboard/layout.tsx
@@ -2,13 +2,11 @@ import { PageHeaderSlot, PageHeaderSlotProvider } from '@/components/ui/page-hea
 import { PageContent, PageHeader, PageLayout } from '@/components/ui/page-layout';
 import { ResizablePanel, ResizablePanelGroup } from '@/components/ui/resizable';
 
-import { GlobalDialogs } from './_components/global-dialogs';
 import { Navbar } from './_components/navbar';
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
   return (
     <>
-      <GlobalDialogs />
       <ResizablePanelGroup direction="horizontal" className="h-full min-h-screen items-stretch">
         <Navbar />
         <ResizablePanel id="main-panel" defaultSize={80}>
