+++ src/components/cards/accounts-card/add-transaction-menu-item.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useCreateTransactionDialog } from '@/components/providers/create-transaction-dialog-provider';
+import { useCreateTransactionDialog } from '@/components/dialogs/create-transaction-dialog';
 import { DropdownMenuItem } from '@/components/ui/dropdown-menu';
 
 interface AddTransactionMenuItemProps {
