+++ src/components/dialogs/create-transaction-dialog/index.ts
@@ -1,1 +1,5 @@
 export { CreateTransactionDialog } from './create-transaction-dialog';
+export {
+  CreateTransactionDialogProvider,
+  useCreateTransactionDialog,
+} from './create-transaction-dialog-context';
