+++ src/components/dialogs/update-transaction-dialog/index.ts
@@ -1,1 +1,5 @@
 export { UpdateTransactionDialog } from './update-transaction-dialog';
+export {
+  UpdateTransactionDialogProvider,
+  useUpdateTransactionDialog,
+} from './update-transaction-dialog-context';
