+++ src/components/dialogs/upsert-account-dialog/index.ts
@@ -1,1 +1,5 @@
 export { UpsertAccountDialog } from './upsert-account-dialog';
+export {
+  UpsertAccountDialogProvider,
+  useUpsertAccountDialog,
+} from './upsert-account-dialog-context';
